+++ app.js
@@ -9,7 +9,8 @@ require("./db.js");
 
 var indexRouter = require("./routes/api/v1/index");
 var usersRouter = require("./routes/api/v1/users");
-var loginRouter = require("./routes/api/v1/login")
+var loginRouter = require("./routes/api/v1/login");
+var registerRouter = require("./routes/api/v1/register.js");
 
 var app = express();
 
@@ -25,7 +26,8 @@ app.use(express.static(path.join(__dirname, "public")));
 
 app.use("/api/v1", indexRouter);
 app.use("/api/v1/users", usersRouter);
-app.use("/api/v1/login", loginRouter)
+app.use("/api/v1/login", loginRouter);
+app.use("/api/v1/register", registerRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
